JACK Audio recording now works!
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@2382 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/SoundFile.cpp
@@ -87,7 +87,7 @@ SoundFile::getLittleEndianFromInteger(unsigned int value, unsigned int length)
 
     do
     {
-        r += (unsigned char)((value >> (8 * r.length())) & 0xff);
+        r += (unsigned char)((long)((value >> (8 * r.length())) & 0xff));
     }
     while (r.length() < length);
